Reduce the number of exceptions thrown.

+++ Development/GeoTransformer/AutoUpdater.cs
@@ -185,7 +185,8 @@ namespace GeoTransformer
 
                     try
                     {
-                        Directory.Delete("PendingUpdate", true);
+                        if (System.IO.Directory.Exists("PendingUpdate"))
+                            Directory.Delete("PendingUpdate", true);
                     }
                     catch (IOException)
                     {
